+++ src/Exception/NumverifyApiResponseException.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * This file is part of the Numverify API Client for PHP.
  *
  * (c) 2024 Eric Sizemore <admin@secondversion.com>
- * (c) 2018-2021 Mark Rogoyski
+ * (c) 2018-2021 Mark Rogoyski <mark@rogoyski.com>
  *
  * @license The MIT License
  *
@@ -17,15 +17,8 @@ declare(strict_types=1);
 namespace Numverify\Exception;
 
 use RuntimeException;
-use stdClass;
 
 /**
  * Thrown when the Numverify API returns an API response that is unexpected.
  */
-class NumverifyApiResponseException extends RuntimeException
-{
-    public function __construct(string $message)
-    {
-        parent::__construct($message);
-    }
-}
+class NumverifyApiResponseException extends RuntimeException {}