+++ src/PhoneNumber/PhoneNumberInterface.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * This file is part of the Numverify API Client for PHP.
  *
  * (c) 2024 Eric Sizemore <admin@secondversion.com>
- * (c) 2018-2021 Mark Rogoyski
+ * (c) 2018-2021 Mark Rogoyski <mark@rogoyski.com>
  *
  * @license The MIT License
  *
@@ -16,12 +16,21 @@ declare(strict_types=1);
 
 namespace Numverify\PhoneNumber;
 
+use JsonSerializable;
+use Stringable;
+
 /**
  * Interface for all phone numbers returned from the Numverify validate API.
  */
-interface PhoneNumberInterface
+interface PhoneNumberInterface extends JsonSerializable, Stringable
 {
+    /**
+     * Is the phone number valid?
+     */
     public function isValid(): bool;
 
+    /**
+     * Get the phone number.
+     */
     public function getNumber(): string;
 }